+++ src/autoscaling/autoscaled_pool.js
@@ -51,10 +51,13 @@ const DEFAULT_OPTIONS = {
  *         // Run some resource-intensive asynchronous operation here.
  *     },
  *     isTaskReadyFunction: async () => {
- *         // Tell the pool whether more tasks are ready to be processed. (true / false)
+ *         // Tell the pool whether more tasks are ready to be processed.
+ *         // Return true or false
  *     },
  *     isFinishedFunction: async () => {
- *         // Tell the pool whether it should finish or wait for more tasks to become available. (true / false)
+ *         // Tell the pool whether it should finish
+ *         // or wait for more tasks to become available.
+ *         // Return true or false
  *     }
  * });
  *
@@ -370,7 +373,7 @@ export default class AutoscaledPool {
      */
     async _maybeFinish() {
         if (this.queryingIsFinished) return;
-        if (this.runningCount > 0) return;
+        if (this.currentConcurrency > 0) return;
 
         this.queryingIsFinished = true;
         try {