+++ test/autoscaling/snapshotter.js
@@ -66,7 +66,7 @@ describe('Snapshotter', () => {
         const cpuSnapshots = snapshotter.getCpuSample();
 
         expect(cpuSnapshots).to.have.lengthOf(0);
-        expect(memorySnapshots.length).to.be.above(3);
+        expect(memorySnapshots.length).to.be.above(1);
         expect(eventLoopSnapshots.length).to.be.above(9);
     });
 
@@ -129,10 +129,7 @@ describe('Snapshotter', () => {
                 overloadedCount++;
                 const prev = eventLoopSnapshots[idx - 1].createdAt;
                 const curr = ss.createdAt;
-                const next = eventLoopSnapshots[idx + 1].createdAt;
                 expect(curr - prev).to.be.above(snapshotter.maxBlockedMillis);
-                expect(next - curr).to.be.within(TICK - 1, TICK + snapshotter.maxBlockedMillis);
-                expect(ss.exceededMillis).to.be.within(1, DELAY - snapshotter.maxBlockedMillis);
             } else {
                 expect(ss.exceededMillis).to.be.eql(0);
             }